# Models/PromptModel.cs
namespace Stressless_Service.Models
{
    public class PromptModel
    {
        public int ID { get; set; }
        public string Type { get; set; }
        public string Text { get; set; }
    }
}