# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "StressLess API",
        "version": "v1"
    },
    "paths": {
        "/Data/GetPrompt": {
            "get": {
                "tags": [
                    "Data"
                ],
                "parameters": [
                    {
                        "name": "PromptID",
                        "in": "query",
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/PromptModel"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PromptModel"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PromptModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/Data/InsertPrompt": {
            "post": {
                "tags": [
                    "Data"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PromptModel"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PromptModel"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/PromptModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/Data/GetConfiguration": {
            "get": {
                "tags": [
                    "Data"
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConfigurationModel"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConfigurationModel"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConfigurationModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/Data/InsertConfiguration": {
            "post": {
                "tags": [
                    "Data"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfigurationModel"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfigurationModel"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfigurationModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/WeatherForecast": {
            "get": {
                "tags": [
                    "WeatherForecast"
                ],
                "operationId": "GetWeatherForecast",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WeatherForecast"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WeatherForecast"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WeatherForecast"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CalenderModel": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "location": {
                        "type": "string",
                        "nullable": true
                    },
                    "start_time": {
                        "type": "string",
                        "nullable": true
                    },
                    "end_time": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "ConfigurationModel": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "firstname": {
                        "type": "string",
                        "nullable": true
                    },
                    "lastname": {
                        "type": "string",
                        "nullable": true
                    },
                    "workingDays": {
                        "type": "string",
                        "nullable": true
                    },
                    "start_time": {
                        "type": "string",
                        "nullable": true
                    },
                    "finish_time": {
                        "type": "string",
                        "nullable": true
                    },
                    "calenderImport": {
                        "type": "string",
                        "nullable": true
                    },
                    "calender": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CalenderModel"
                        },
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "PromptModel": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "text": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "WeatherForecast": {
                "type": "object",
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "temperatureC": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "temperatureF": {
                        "type": "integer",
                        "format": "int32",
                        "readOnly": true
                    },
                    "summary": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            }
        }
    }
}